Split scripts by service
