+++ src/v8/admin/after-install/add-dialog-content-to-edit-dialog.ts
