+++ src/v8/admin/after-install/comet-config-provider.ts
