+++ src/v8/admin/after-install/merge-admin-theme-into-admin.ts
