+++ src/v8/admin/after-install/merge-blocks-admin-into-cms-admin.ts
