+++ src/v8/admin/after-install/mui-x-codemods.ts
@@ -1,4 +1,4 @@
-import { executeCommand } from "../util/execute-command.util";
+import { executeCommand } from "../../../util/execute-command.util";
 
 export default async function executeMuiXCodemods() {
     await executeCommand("npx", ["@mui/x-codemod@latest", "v6.0.0/data-grid/preset-safe", "admin/src"]);
