+++ src/v8/admin/after-install/rename-menu-components-in-admin.ts
