+++ src/v8/admin/after-install/tooltip-2-remove-trigger-prop.ts
