+++ src/v8/admin/before-install/remove-blocks-packages-admin.ts
@@ -1,0 +1,13 @@
+import { existsSync } from "fs";
+
+import { PackageJson } from "../../../util/package-json.util";
+
+export const stage = "before-install";
+
+export default async function removeBlocksPackagesAdmin() {
+    if (existsSync("admin/package.json")) {
+        const packageJson = new PackageJson("admin/package.json");
+        packageJson.removeDependency("@comet/blocks-admin");
+        packageJson.save();
+    }
+}
