+++ src/v8/admin/before-install/update-react-dependencies.ts
@@ -1,6 +1,6 @@
 import { existsSync } from "fs";
 
-import { PackageJson } from "../util/package-json.util";
+import { PackageJson } from "../../../util/package-json.util";
 
 export const stage = "before-install";
