+++ src/v8/admin/before-install/use-mui-date-picker-in-grid.ts
@@ -1,7 +1,7 @@
 import { existsSync } from "node:fs";
 import { Project, SyntaxKind } from "ts-morph";
 
-import { PackageJson } from "../util/package-json.util";
+import { PackageJson } from "../../../util/package-json.util";
 
 export const stage = "before-install";
