+++ src/v8/api/after-install/merge-blocks-api-into-cms-api.ts
