+++ src/v8/api/after-install/mikro-orm-base-entity-generic.ts
