+++ src/v8/api/after-install/mikro-orm-delete-rule.ts
