+++ src/v8/api/after-install/remove-current-user-from-auth-module.ts
