+++ src/v8/api/after-install/update-s3-config.ts
