+++ src/v8/api/before-install/move-maxSrcResolution-in-comet-config.ts
