+++ src/v8/api/before-install/replace-nestjs-console-with-nest-commander.ts
@@ -1,6 +1,6 @@
 import { existsSync } from "node:fs";
 
-import { PackageJson } from "../util/package-json.util";
+import { PackageJson } from "../../../util/package-json.util";
 
 export const stage = "before-install";
