+++ src/v8/api/before-install/update-sentry.ts
@@ -1,7 +1,7 @@
 import { existsSync } from "fs";
 import { Project, SyntaxKind } from "ts-morph";
 
-import { PackageJson } from "../util/package-json.util";
+import { PackageJson } from "../../../util/package-json.util";
 
 export const stage = "before-install";
