+++ src/v8/misc/before-install/replace-node-with-v22-locally.ts
@@ -1,7 +1,7 @@
 import { writeFile } from "fs/promises";
 import { existsSync } from "node:fs";
 
-import { PackageJson } from "../util/package-json.util";
+import { PackageJson } from "../../../util/package-json.util";
 
 export const stage = "before-install";
