+++ src/v8/site/after-install/remove-graphql-client-from-site-preview-handlers.ts
